requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
toml>=0.10.0
RPi.GPIO>=0.7.0
adafruit-circuitpython-mpu6050
opencv-python
picamera2
numpy